Use 'osv' environment variable to find Vision source and libraries
The vxa-node module is dependent on Vision source and library files.  As a
first step on the path to 'npm' management of its build and installation,
the 'binding.gyp' file that builds it now finds the source and libraries it
needs using the value of the 'osv' environment variable.  That variable must
be set to must name the vision software source directory containing an already
built version of at least the Vision libraries:

   export osv=$HOME/Vision/vision/sofware/src/master/src
   cd $osv
   visionBuilder

Once these steps have been run, 'vxanode' can be consumed by a node module by
adding a reference to its 'github' repository (currently personal and private)
to the consuming module's 'package.json' 'dependencies':

   ...
   "dependencies": {
      ...
      "vxanode": "MichaelJCaruso/vxa-node",
      ...
   }

Once that's done, 'npm install' should be able to automatically fetch, build,
install it for use by the consuming module.