Refactor and semi-package vxa server (vxa.o)
Among other things, this code now supports a state change callback useful
for detecting 'stop' events.  Originally intended for stopping node after
all vxa services have stopped, it is not sufficient.  The 'node' process
stays alive as long as at least one active 'uv_async_t' handle exists.
Because 'vxa' creates and keeps one of those around as soon it evaluates
a query or offers a service, 'node's event loop is kept alive and running
indefinitely.  To prevent that, 'Vxa' must deactivate that handle as soon
as it is no longer needed. In practice that means no 'VA::Node::Export's.
Reliably detecting that may turn out to be interesting in the presence of
failed peers (doable, but interesting).